Update my-element.js
This is the base class for extending lit-element. However we need to re-install node in our repo because it doesn't have all the packages (lit-element is missing for example, therefor this class won't work at the moment). I tried to do it myself in my own branch but I've never used branches and when I tried to push my changes with the node folder to this branch it didn't work. I looked at how to do it but still didn't manage. I will ask for help tomorrow about it.